Add files via upload
In this project, I aimed to build a Diabetes Prediction Model using a Support Vector Machine (SVM) classifier.

The process began with exploring and understanding the dataset, which included features such as BMI, blood glucose levels, insulin levels, and other health-related indicators. After accessing the diabetes dataset, I performed data preprocessing, which included standardizing the feature values to ensure consistency.

The dataset was then split into training and testing sets, and the SVM model was trained on the training data. After training, I evaluated the model's performance using accuracy and other relevant metrics.

The final goal of this project is to enable the model to predict whether a person is diabetic or not when provided with new input data.